third_party/WebKit/LayoutTests/inspector/sources/debugger-async/async-callstack-middle-run.html

<html>
<head>
<script src="../../../http/tests/inspector/inspector-test.js"></script>
<script src="../../../http/tests/inspector/debugger-test.js"></script>
<script>

function testFunction()
{
    setTimeout(timeoutOffCapturing, 0);
}

function timeoutOffCapturing()
{
    setTimeout(timeoutOffCapturing2, 0);
    debugger;
    setTimeout(timeoutOnCapturing, 0);
}

function timeoutOffCapturing2()
{
    debugger;
}

function timeoutOnCapturing()
{
    debugger;
}

var test = function() {
  var totalDebuggerStatements = 3;
  var maxAsyncCallStackDepth = 8;

  SourcesTestRunner.setQuiet(true);
  SourcesTestRunner.startDebuggerTest(step1);

  async function step1() {
    await TestRunner.DebuggerAgent.setAsyncCallStackDepth(0);
    SourcesTestRunner.runTestFunctionAndWaitUntilPaused(didPause);
  }

  function resumeExecution() {
    SourcesTestRunner.resumeExecution(SourcesTestRunner.waitUntilPaused.bind(InspectorTest, didPause));
  }

  var step = 0;
  var callStacksOutput = [];
  function didPause(callFrames, reason, breakpointIds, asyncStackTrace) {
    ++step;
    if (step === 1) {
      TestRunner.DebuggerAgent.setAsyncCallStackDepth(maxAsyncCallStackDepth).then(resumeExecution);
      return;
    }

    callStacksOutput.push(SourcesTestRunner.captureStackTraceIntoString(callFrames, asyncStackTrace) + '\n');
    if (step < totalDebuggerStatements) {
      resumeExecution();
    } else {
      TestRunner.addResult('Captured call stacks in no particular order:');
      callStacksOutput.sort();
      TestRunner.addResults(callStacksOutput);
      SourcesTestRunner.completeDebuggerTest();
    }
  }
};

</script>
</head>

<body onload="runTest()">
<p>
Tests that capturing asynchronous call stacks in debugger works if started after some time since the page loads.
</p>

</body>
</html>
