third_party/WebKit/LayoutTests/inspector/sources/debugger-async/async-callstack-middle-run-expected.txt

Tests that capturing asynchronous call stacks in debugger works if started after some time since the page loads.

Set timer for test function.
Captured call stacks in no particular order:
Call stack:
    0) timeoutOffCapturing2 (async-callstack-middle-run.html:21)

Call stack:
    0) timeoutOnCapturing (async-callstack-middle-run.html:26)
    [setTimeout]
    0) timeoutOffCapturing (async-callstack-middle-run.html:16)
